Refactor Command.cpp separate into each cmd.cpp
