+++ Makefile
@@ -36,7 +36,16 @@ SUBFILE1_SRC = server.cpp
 SUBFILE2_SRC = autosocket.cpp
 SUBFILE3_SRC = parser.cpp
 SUBFILE4_SRC = user.cpp
-SUBFILE5_SRC = command.cpp
+SUBFILE5_SRC = command.cpp \
+			   invite.cpp \
+			   join.cpp \
+			   kick.cpp \
+			   nick.cpp \
+			   pong.cpp \
+			   privmsg.cpp \
+			   quit.cpp \
+			   topic.cpp \
+			   user.cpp
 SUBFILE6_SRC = channel.cpp
 
 SRC =	main.cpp	\
