+++ srcs/command/user.cpp
@@ -1,0 +1,27 @@
+#include "command.hpp"
+#include "reply.hpp"
+
+//  The USER message is used at the beginning of connection to specify
+//    the username, hostname, servername and realname of s new user.
+void  cmd::username(server &svr, int poll_fd_pos, std::string str) {
+  poll_fd pollfd = svr.get_pollfd(poll_fd_pos);
+  user &usr = svr.get_user(pollfd.fd);
+
+  // Separar el resto del realname
+  std::vector<std::string> cmd_params = ft_split(str, ':');
+  std::vector<std::string> other_params = ft_split(cmd_params[0], ' ');
+  if (other_params.size() + 1 < 4)
+  {
+    svr.send_message(": 461 USER : <command> :Not enough parameters \r\n", usr.get_fd());
+    return ;
+  }
+  usr.set_username(other_params[0]);
+  usr.set_hostname(other_params[1]);
+  usr.set_servername(other_params[2]);
+  usr.set_realname(cmd_params[1]); //TODO sus
+
+  // IF registered --> Send RPL_WELCOME
+  //TODO meterlo en el define
+  usr.is_registered(svr);
+  std::cout << usr << std::endl;
+}
